requirements.txt
flask==2.3.3
transformers==4.34.0
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
newspaper3k==0.2.8
nltk==3.8.1
textblob==0.17.1
beautifulsoup4==4.12.2
requests==2.31.0
plotly==5.15.0
wordcloud==1.9.2
spacy==3.7.0
python-dotenv==1.0.0
gunicorn==21.2.0